+++ hexital/utils/__init__.py
@@ -1,2 +1,2 @@
-from . import candlesticks, timeframe, indexing  # noqa F401
+from . import candles, timeframe, indexing  # noqa F401
 from .timeframe import TimeFrame