+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 - Changed Sub/managed indicators to auto populate candles field
 - Changed 'as_list' property a method that can now take a nested indicator name
 - Removed read property
+- Renamed utils/candlesticks to utils/candles
 
 ## 0.4.0 - 2024-01-22
 - Cleaned up code and some potential Bugs ruff/pyright